separate input.json from cube_data.json to keep human readability
