+++ python/mrchem/input_parser/README.md
@@ -1,2 +1,2 @@
-This file was automatically generated by parselglossy on 2021-06-17
+This file was automatically generated by parselglossy on 2021-06-21
 Editing is *STRONGLY DISCOURAGED*
