+++ python/mrchem/input_parser/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-# This file was automatically generated by parselglossy on 2021-06-17
+# This file was automatically generated by parselglossy on 2021-06-21
 # Editing is *STRONGLY DISCOURAGED*
