+++ python/mrchem/input_parser/api.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# This file was automatically generated by parselglossy on 2021-06-17
+# This file was automatically generated by parselglossy on 2021-06-21
 # Editing is *STRONGLY DISCOURAGED*
 
 from copy import deepcopy
