+++ python/mrchem/input_parser/plumbing/lexer.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# This file was automatically generated by parselglossy on 2021-06-17
+# This file was automatically generated by parselglossy on 2021-06-21
 # Editing is *STRONGLY DISCOURAGED*
 
 import json
